fixed audio export [4.7]

+++ src/project/internal/exportprojectscenario.cpp
@@ -407,6 +407,11 @@ Ret ExportProjectScenario::doExportLoop(const muse::io::path_t& scorePath, std::
     }
 
     while (true) {
+        //! NOTE Most writers write data to a given device (buffer)
+        //! But there are two atypical cases:
+        //! 1. Audio export - writes directly to a file at the specified path
+        //! 2. Export score to unpacked directory - creates a directory and writes files to it
+
         Buffer outputBuf;
         outputBuf.setMeta("file_path", scorePath.toStdString());
         IF_ASSERT_FAILED(outputBuf.open(IODevice::WriteOnly)) {
@@ -415,8 +420,14 @@ Ret ExportProjectScenario::doExportLoop(const muse::io::path_t& scorePath, std::
 
         Ret ret = exportFunction(outputBuf);
         outputBuf.close();
+
+        const bool isFileMode = fileSystem()->exists(scorePath);
+
         if (!ret) {
             if (ret.code() == static_cast<int>(Ret::Code::Cancel)) {
+                if (isFileMode) {
+                    fileSystem()->remove(scorePath);
+                }
                 return ret;
             }
 
@@ -427,12 +438,14 @@ Ret ExportProjectScenario::doExportLoop(const muse::io::path_t& scorePath, std::
             }
         }
 
-        ret = fileSystem()->writeFile(scorePath, outputBuf.data());
-        if (!ret) {
-            if (askForRetry(filename)) {
-                continue;
-            } else {
-                return make_ret(Ret::Code::Cancel);
+        if (!isFileMode) {
+            ret = fileSystem()->writeFile(scorePath, outputBuf.data());
+            if (!ret) {
+                if (askForRetry(filename)) {
+                    continue;
+                } else {
+                    return make_ret(Ret::Code::Cancel);
+                }
             }
         }
